Extract parser DML grammar
